fix: Course and Workshop Page
