+++ app/chef/course/page.tsx
@@ -1,11 +1,11 @@
 import Navbar from '../components/Navbar'
-import VideoTable from '../components/video/VideoTable'
+import CourseTable from '../components/course/CourseTable'
 
 export default function Home() {
   return (
     <div className="animate-fadeIn">
       <Navbar />
-      <VideoTable />
+      <CourseTable />
     </div>
   )
 }
